+++ CHANGELOG.md
@@ -2,12 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Breaking
+
+- Dropped support for versions of Go before 1.13
+
 
 ## v0.14.0 — 2019-08-08
 
 ### Changed
 
-- Only return the token from `Token` once, then return `io.EOF`.
+- Only return the token from `Token` once, then return `io.EOF`
 
 
 ## v0.13.6 — 2019-08-06